requirements.txt
bcrypt==4.1.2
cffi==1.16.0
cryptography==41.0.7
paramiko==3.4.0
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
